nanite qdel fix

+++ code/modules/mob/living/carbon/superior_animal/roach/types/nanites.dm
@@ -97,12 +97,13 @@
 	..()
 	parent = nuparent
 
-/mob/living/simple_animal/hostile/naniteswarm/death()
+/mob/living/simple_animal/hostile/naniteswarm/death(gibbed)
 	..()
 	if(parent)
 		parent.nanite_swarms.Remove(src)
 	new /obj/effect/decal/cleanable/blood/oil(get_turf(src))
-	qdel(src)
+	if(!gibbed)		// Parent roach gibs all child nanites on death, which qdels the nanites.
+		qdel(src)
 
 /mob/living/simple_animal/hostile/naniteswarm/Destroy()
 	if(parent)
